Moved one_hot function to the functions script.
